backend/requirements.txt
flask
flask-cors
opencv-python
pytesseract
scikit-learn==1.6.1
tensorflow
pandas
numpy
statsmodels
supabase
Pillow
joblib
python-dotenv
